requirements-test.txt
behave==1.2.6
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
mock==4.0.2
neo4j==4.0.0
parse==1.15.0
parse-type==0.5.2
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.15.0
sure==1.4.11
urllib3==1.25.9
